pyproject.toml
[tool.poetry]
name = "ca-ai-template"
version = "0.1.0"
description = "A simple FastAPI project"
authors = ["kate"]

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.74.0"
uvicorn = "^0.15.0"
python-dotenv = "^1.0.0"
alembic = "^1.10.2"
aiomysql = "^0.0.22"
greenlet = "^1.1.2"
pymysql = "^0.9.3"
cryptography = "^40.0.1"
requests = "^2.28.2"
jinja2 = "^3.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
